feat(playground): refactor bricks to encapsulate drag logic
