+++ modules/code-builder/playground/pages/WorkSpace/index.tsx
@@ -1,60 +1,11 @@
-import {
-  BrickBlock,
-  BrickData,
-  BrickExpression,
-  BrickStatement,
-  ModelBrickBlock,
-  ModelBrickData,
-  ModelBrickExpression,
-  ModelBrickStatement,
-} from '@/brick';
+import BrickFactory from './BrickFactory';
 import { WORKSPACES_DATA } from './data';
 import type { Brick } from './data';
-
-function getBrick(brickData: Brick) {
-  switch (brickData.type) {
-    case 'data':
-      return (
-        <BrickData
-          id={brickData.id}
-          instance={brickData.instance as ModelBrickData}
-          coords={brickData.coords}
-        />
-      );
-    case 'expression':
-      return (
-        <BrickExpression
-          id={brickData.id}
-          instance={brickData.instance as ModelBrickExpression}
-          coords={brickData.coords}
-        />
-      );
-    case 'statement':
-      return (
-        <BrickStatement
-          id={brickData.id}
-          instance={brickData.instance as ModelBrickStatement}
-          coords={brickData.coords}
-        />
-      );
-    case 'block':
-      return (
-        <BrickBlock
-          id={brickData.id}
-          brickData={brickData}
-          instance={brickData.instance as ModelBrickBlock}
-          coords={brickData.coords}
-        />
-      );
-    default:
-      return <></>;
-  }
-}
 
 function RenderBricks({ brickData }: { brickData: Brick }) {
   return (
     <>
-      {getBrick(brickData)}
+      <BrickFactory brickData={brickData} />
       {brickData.children &&
         brickData.children?.length > 0 &&
         brickData.children.map((child) => <RenderBricks key={child.id} brickData={child} />)}
